isolated the dropdownAjax functions into a separate module
