+++ footer.html
@@ -146,6 +146,7 @@ window.ELMO_FEATURES = {
 <script src="node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
 <!-- Dropdown helpers used by select.js -->
 <script src="js/dropdownUtils.js"></script>
+<script src="js/dropdownAjax.js"></script>
 <!-- select-Funktionen laden -->
 <script src="js/select.js"></script>
 <!--Tagify-->
