+++ js/select.js
@@ -6,447 +6,30 @@ const { updateDropdownPlaceholders, filterDataByGEM } =
         filterDataByGEM: window.filterDataByGEM,
       };
 
-/**
- * Fills the timezone dropdown and sets the default timezone based on system settings and user's location
- * @async
- * @function setupTimezoneDropdownAjax
- * @param {string|jQuery|HTMLElement} dropdownSelector - The selector for the timezone dropdown element
- * @param {string} jsonPath - Path to the timezones JSON file
- * @returns {Promise<void>}
- */
-async function setupTimezoneDropdownAjax(dropdownSelector = '#input-stc-timezone', jsonPath = 'json/timezones.json') {
-  try {
-    const $dropdown = $(dropdownSelector);
-    if ($dropdown.length === 0) return;
-
-    /**
-     * Gets system timezone from browser settings
-     * @param {jQuery} $select - The jQuery select element
-     * @returns {string} Timezone offset in format "+HH:MM" or "-HH:MM"
-     */
-    function getSystemTimezone($select) {
-      try {
-        const timezoneName = Intl.DateTimeFormat().resolvedOptions().timeZone;
-
-        if (timezoneName) {
-          const options = $select.find('option').get();
-          const date = new Date();
-          const offset = -date.getTimezoneOffset();
-          const hours = Math.floor(Math.abs(offset) / 60).toString().padStart(2, '0');
-          const minutes = (Math.abs(offset) % 60).toString().padStart(2, '0');
-          const offsetStr = `${offset >= 0 ? '+' : '-'}${hours}:${minutes}`;
-
-          let bestMatch = null;
-
-          for (const option of options) {
-            const optionText = $(option).text();
-            const optionValue = $(option).val();
-
-            if (optionText.includes(`(${timezoneName})`)) {
-              return optionValue;
-            }
-
-            if (optionValue === offsetStr && optionText.includes(timezoneName.split('/')[0])) {
-              bestMatch = optionValue;
-              break;
-            }
-
-            if (optionValue === offsetStr && !bestMatch) {
-              bestMatch = optionValue;
-            }
-          }
-
-          if (bestMatch) return bestMatch;
-        }
-
-        const date = new Date();
-        const offset = -date.getTimezoneOffset();
-        const hours = Math.floor(Math.abs(offset) / 60).toString().padStart(2, '0');
-        const minutes = (Math.abs(offset) % 60).toString().padStart(2, '0');
-        return `${offset >= 0 ? '+' : '-'}${hours}:${minutes}`;
-
-      } catch (error) {
-        console.error('Error getting system timezone:', error);
-        return null;
-      }
-    }
-
-    if ($dropdown.find('option').length > 0) {
-      const systemTimezone = getSystemTimezone($dropdown);
-      setTimezoneInDropdown($dropdown, systemTimezone);
-      return;
-    }
-
-    const response = await fetch(jsonPath);
-    const timezones = await response.json();
-
-    /**
-     * Extracts UTC offset from timezone label
-     * @param {string} label - The timezone label (e.g., "UTC+00:00 (Africa/Abidjan)")
-     * @returns {string} The UTC offset (e.g., "+00:00")
-     */
-    function extractUTCOffset(label) {
-      const match = label.match(/UTC([+-]\d{2}:\d{2})/);
-      return match ? match[1] : '';
-    }
-
-    $dropdown.empty();
-    timezones.forEach(timezone => {
-      $dropdown.append(
-        $('<option>', {
-          value: extractUTCOffset(timezone.label),
-          text: timezone.label
-        })
-      );
-    });
-
-    /**
-     * Sets the dropdown value to the specified timezone option
-     * @param {jQuery} $select - The jQuery select element
-     * @param {string} timezoneName - The timezone name (e.g., "Europe/Berlin")
-     * @returns {boolean} True if timezone was set successfully
-     */
-    function setTimezoneInDropdown($select, timezoneName) {
-      if (!timezoneName) return false;
-
-      const allOptions = Array.from($select.find('option'));
-      const exactMatch = allOptions.find(option =>
-        option.text.includes(`(${timezoneName})`)
-      );
-
-      if (exactMatch) {
-        $select.find('option').prop('selected', false);
-        $(exactMatch).prop('selected', true);
-        return true;
-      }
-
-      const region = timezoneName.split('/')[0];
-      const regionMatch = allOptions.find(option =>
-        option.text.includes(`(${region}/`)
-      );
-
-      if (regionMatch) {
-        $select.find('option').prop('selected', false);
-        $(regionMatch).prop('selected', true);
-        return true;
-      }
-
-      return false;
-    }
-
-    const browserTimezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-    if (browserTimezone) {
-      setTimezoneInDropdown($dropdown, browserTimezone);
-    }
-
-  } catch (error) {
-    console.error('Error initializing timezone dropdown:', error);
-  }
-}
+const {
+  setupTimezoneDropdownAjax,
+  setupResourceTypeDropdownAjax,
+  setupLanguageDropdownAjax,
+  setupTitleTypeDropdownAjax,
+  setupLicenseDropdown,
+  addPlaceholder,
+  runSequentialFallback,
+} =
+  (typeof require === 'function' && typeof module !== 'undefined')
+    ? require('./dropdownAjax.js')
+    : {
+        setupTimezoneDropdownAjax: window.setupTimezoneDropdownAjax,
+        setupResourceTypeDropdownAjax: window.setupResourceTypeDropdownAjax,
+        setupLanguageDropdownAjax: window.setupLanguageDropdownAjax,
+        setupTitleTypeDropdownAjax: window.setupTitleTypeDropdownAjax,
+        setupLicenseDropdown: window.setupLicenseDropdown,
+        addPlaceholder: window.addPlaceholder,
+        runSequentialFallback: window.runSequentialFallback,
+      };
 
 /**
  * This script handles the setup and initialization of various dropdowns, event listeners, and autocomplete functions for the metadata editor.
  */
-
-// Dropdown helper functions exposed globally so tests can invoke them
-function setupResourceTypeDropdownAjax() {
-  const select = $("#input-resourceinformation-resourcetype");
-  if (select.length === 0) return;
-
-  select.prop('disabled', true).empty().append(
-    $("<option>", {
-      value: "",
-      text: "Loading...",
-    })
-  );
-
-  $.ajax({
-    url: "api/v2/vocabs/resourcetypes",
-    method: "GET",
-    dataType: "json",
-    success: function (data) {
-      select.empty();
-      addPlaceholder(select, true);
-
-      if (Array.isArray(data)) {
-        const isGEM = window.ELMO_FEATURES?.showGGMsProperties;
-        const filteredData = filterDataByGEM(data, 'resourceType', isGEM);
-
-        filteredData.forEach(function (type) {
-          select.append(
-            $("<option>", {
-              value: type.id,
-              text: type.resource_type_general,
-              title: type.description,
-            })
-          );
-        });
-      }
-    },
-    error: function (jqXHR, textStatus, errorThrown) {
-      console.error("Error loading resource types:", textStatus, errorThrown);
-      select.empty().append(
-        $("<option>", {
-          value: "",
-          text: "Error loading data",
-        })
-      );
-    },
-    complete: function () {
-      select.prop('disabled', false).trigger("change");
-    },
-  });
-}
-
-function setupLanguageDropdownAjax() {
-  const select = $("#input-resourceinformation-language");
-  if (select.length === 0) return;
-
-  select.prop('disabled', true).empty().append(
-    $("<option>", {
-      value: "",
-      text: "Loading...",
-    })
-  );
-
-  $.ajax({
-    url: "api/v2/vocabs/languages",
-    method: "GET",
-    dataType: "json",
-    success: function (data) {
-      select.empty();
-      addPlaceholder(select, true);
-
-      if (Array.isArray(data)) {
-        const isGEM = window.ELMO_FEATURES?.showGGMsProperties;
-        const filteredData = filterDataByGEM(data, 'language', isGEM);
-
-        filteredData.forEach(function (lang) {
-          select.append(
-            $("<option>", {
-              value: lang.id,
-              text: lang.name,
-              title: lang.code,
-            })
-          );
-        });
-
-        // Pre-select English (code "en") as default
-        const englishOption = filteredData.find(lang => lang.code === 'en');
-        if (englishOption) {
-          select.val(englishOption.id);
-        }
-      }
-    },
-    error: function (jqXHR, textStatus, errorThrown) {
-      console.error("Error loading languages:", textStatus, errorThrown);
-      select.empty().append(
-        $("<option>", {
-          value: "",
-          text: "Error loading data",
-        })
-      );
-    },
-    complete: function () {
-      select.prop('disabled', false);
-    },
-  });
-}
-// This function is a fallback implementation for the environments that don't have fetch in it.
-function setupTitleTypeDropdownAjax() {
-  const select = $("#input-resourceinformation-titletype");
-  if (select.length === 0) {
-    console.error("Title type dropdown not found. Ensure the element with ID 'input-resourceinformation-titletype' exists in the DOM.");
-    return;
-  }
-
-  select.prop('disabled', true).empty().append(
-    $("<option>", {
-      value: "",
-      text: "Loading...",
-    })
-  );
-
-  $.ajax({
-    url: "api/v2/vocabs/titletypes",
-    method: "GET",
-    dataType: "json",
-    success: function (data) {
-      select.empty().append(
-        $("<option>", {
-          value: "",
-          text: "Choose...",
-          "data-translate": "general.choose",
-        })
-      );
-
-      let mainTitleId = "";
-      let alternativeTitleId = "";
-      if (Array.isArray(data)) {
-        data.forEach(function (type) {
-          const option = $("<option>", {
-            value: type.id,
-            text: type.name,
-          });
-
-          select.append(option);
-
-          if (type.name.toLowerCase() === "main title") {
-            mainTitleId = type.id.toString();
-          }
-          if (type.name.toLowerCase() === "alternative title") {
-            alternativeTitleId = type.id.toString();
-          }
-        });
-      }
-
-      if (mainTitleId) {
-        select.val(mainTitleId);
-        window.mainTitleTypeId = mainTitleId;
-      }
-      window.alternativeTitleTypeId = alternativeTitleId || "";
-
-      window.titleTypeOptionsHtml = select.html();
-    },
-    error: function (jqXHR, textStatus, errorThrown) {
-      console.error("Error loading title types:", textStatus, errorThrown);
-      select.empty().append(
-        $("<option>", {
-          value: "",
-          text: "Error loading data",
-        })
-      );
-    },
-    complete: function () {
-      select.prop('disabled', false);
-    },
-  });
-}
-
-/**
-* Populates the select field with ID input-rights-license with options created via an API call.
-* @param {boolean} isSoftware - Determines whether to retrieve licenses for software or all resource types.
-*/
-function setupLicenseDropdown(isSoftware) {
-  const $select = $("#input-rights-license"); // Defined as $select for consistency
-  const top_licenseId = "CC-BY-4.0"; //Should be the first
-  const copyleftLicenses = ['GPL-3.0-or-later', 'EUPL-1.2']; // Should be the last
-
-  // 1. Determine the endpoint FIRST
-  const endpoint = isSoftware ? "vocabs/licenses/software" : "vocabs/licenses/all";
-
-  // Loading state
-  $select.prop("disabled", true).empty().append(
-    $("<option>", {
-      value: "",
-      text: "Loading...",
-      "data-translate": "general.loading",
-    })
-  );
-
-  // 2. Start the API call
-  $.getJSON(`./api/v2/${endpoint}`, function (data) {
-    let processedLicenses = [];
-
-    // Prepare the options for the dropdown menu
-    if (!isSoftware) {
-      // Non-software
-      processedLicenses = data
-        .filter(item => item.forSoftware === "0") // Only non-software
-        .sort((a, b) => {
-          // Custom Priority: If it's our target ID, move it to the top (-1)
-          if (a.rightsIdentifier === top_licenseId) return -1;
-          if (b.rightsIdentifier === top_licenseId) return 1;
-
-          // Otherwise: Standard alphabetical sort
-          return a.rightsIdentifier.localeCompare(b.rightsIdentifier);
-        });
-    } else {
-      // Software
-      processedLicenses = data
-        .filter(item => item.forSoftware === "1") // Only software licenses
-        .sort((a, b) => {
-          const aIsCopyleft = copyleftLicenses.includes(a.rightsIdentifier);
-          const bIsCopyleft = copyleftLicenses.includes(b.rightsIdentifier);
-          
-          if (aIsCopyleft !== bIsCopyleft) {
-            return aIsCopyleft ? 1 : -1; // Non-copyleft first
-          }
-          return a.rightsIdentifier.localeCompare(b.rightsIdentifier);
-        });
-    }
-    // Clear existing options
-    $select.empty()
-
-    // Include them into the dropdown
-    processedLicenses.forEach(license => {
-      const option = $("<option>", {
-        value: license.rights_id,
-        text: `${license.text} (${license.rightsIdentifier})`,
-        title: license.description || license.text
-      });
-
-      if (license.rightsIdentifier === "CC-BY-4.0") {
-        option.prop("selected", true);
-      }
-
-      $select.append(option);
-    });
-
-    $select.prop("disabled", false).trigger("change");
-
-  }).fail(function (jqXHR, textStatus, errorThrown) {
-    // Fallback: use CC-BY-4.0 (rights_id=1) if API call fails
-    console.error("Error loading licenses:", textStatus, errorThrown);
-    $select.empty().append(
-      $("<option>", {
-        value: "1",
-        text: "Creative Commons Attribution 4.0 International (CC-BY-4.0)",
-        selected: true
-      })
-    );
-
-    $select.prop("disabled", false).trigger("change");
-  });
-}
-
-/**
- * Adds a "Choose..." placeholder option to a dropdown
- * For ICGEM-specific dropdowns, skips placeholder when ICGEM mode is enabled
- * @param {jQuery} $select - The jQuery select element
- * @param {boolean} isGEMDropdown - Whether this is a ICGEM-specific dropdown (skips placeholder if ICGEM enabled)
- */
-function addPlaceholder($select, isGEMDropdown = false) {
-  const isGEM = window.ELMO_FEATURES?.showGGMsProperties;
-  
-  // For GEM dropdowns, don't add placeholder when GEM is enabled. For others, always add.
-  if (isGEMDropdown && isGEM) return;
-  
-  // Use translated text if translations are already loaded, otherwise fall back to English
-  const translatedText = window.elmo?.translate?.('general.choose') || 'Choose...';
-  
-  $select.append(
-    $("<option>", { value: "", text: translatedText, "data-translate": "general.choose" })
-  );
-}
-
-// Make AJAX fallback dropdown setup functions available globally (important for tests)
-window.setupTimezoneDropdownAjax = setupTimezoneDropdownAjax;
-window.setupLicenseDropdown = setupLicenseDropdown;
-window.setupLanguageDropdownAjax = setupLanguageDropdownAjax;
-window.setupResourceTypeDropdownAjax = setupResourceTypeDropdownAjax;
-window.setupTitleTypeDropdownAjax = setupTitleTypeDropdownAjax;
-
-/**
- * Helper to run legacy sequential fallbacks in one place.
- */
-function runSequentialFallback() {
-  console.warn("Falling back to legacy sequential dropdown initialization.");
-  setupTimezoneDropdownAjax();
-  setupResourceTypeDropdownAjax();
-  setupLanguageDropdownAjax();
-  setupTitleTypeDropdownAjax();
-}
 
 /**
  * Initializes all dropdowns in parallel for faster page load.
@@ -1173,7 +756,9 @@ if (typeof module !== 'undefined' && module.exports) {
     setupResourceTypeDropdownAjax,
     setupLanguageDropdownAjax,
     setupTitleTypeDropdownAjax,
+    setupLicenseDropdown,
     setupIdentifierTypesDropdown,
+    runSequentialFallback,
     populateTimezoneDropdownWithData,
     populateResourceTypeDropdownWithData,
     populateLanguageDropdownWithData,
